requirements.txt
alembic==1.14.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
Flask-Migrate==4.1.0
psycopg2==2.9.10
pytest==8.3.4
marshmallow==3.26.1
PyJWT==2.10.1
Flask-OAuthlib==0.9.6
Authlib==1.4.1
requests==2.32.3
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
PyJWT==2.6.0